[LIBAUTOCORK]: Beter TCP_CORK mimic
Normally apps do TCP_CORK[1] before a series of writes and then a TCP_CORK[0],
we were doing TCP_CORK[0]/[1] in quick sequence.

This was done trying to detect why in some cases the delayed ack timer was
triggering when doing autocorking, but didn't solved this problem.

It turns out that if we do a clock_gettime after setsockopt(TCP_CORK, 0) we
somehow make the delayed acks to trigger, checking now how to avoid this.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>