[LIBAUTOCORK]: Periodically dump the stats
atexit() routines are not called when the host program is terminated with a
signal, so we must periodically dump the stats if we want to see them.

By default do it every 5 minutes, using the timestamp already taken at
push_pending_frames.

The AUTOCORK_DUMP_INTERVAL (seconds) environment variable can be used to tune
the dump interval. Setting it to 0 disables stats collection and dumping.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>